requirements.txt
aniso8601==10.0.0
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
Faker==36.1.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.1.3
